+++ changelog.md
@@ -15,6 +15,7 @@
   - Updated `actions/setup-node` to v2.
   - Don’t specify the `CI` environment variable as it’s set by default.
 - Stop snapshot testing Node.js versions separately as now all supported versions result in the same CLI output.
+- Improved the test helper function `replaceStackTraces`.
 
 ## 1.2.0